# api/openapi.yaml
openapi: 3.1.0
info:
  title: ClinicOS API
  version: 1.0.0
  description: |
    Specialty-First EMR & Practice Management API for Indian Clinics.
    Built by RH Technology, Pune.

    ## Authentication
    Use Bearer token from the `/auth/login` endpoint.
    Include it on every authenticated request:
    ```
    Authorization: Bearer {token}
    ```

    ## Multi-tenancy
    All endpoints are automatically scoped to the authenticated user's clinic.
    No `clinic_id` parameter is needed; it is derived from the token.

    ## Rate Limiting
    100 requests/minute per clinic. Returns HTTP 429 on breach.

    ## Specialties Supported
    `dermatology` | `physiotherapy` | `dental` | `ophthalmology` | `orthopaedics` | `ent` | `gynaecology`

    ## Date / Time Format
    All dates use ISO 8601: `YYYY-MM-DD` for dates, `YYYY-MM-DDTHH:mm:ssZ` for timestamps.

    ## Pagination
    Paginated responses follow a consistent envelope:
    ```json
    { "data": [...], "meta": { "current_page": 1, "per_page": 15, "total": 120, "last_page": 8 } }
    ```
servers:
  - url: https://api.clinicos.in/api/v1
    description: Production

  - url: https://staging-api.clinicos.in/api/v1
    description: Staging
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Sanctum personal access token from /auth/login

  # ───────────────────────────────────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "The given data was invalid."
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            phone: ["The phone field is required."]
        code:
          type: string
          nullable: true
          example: "VALIDATION_FAILED"
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            message: "Unauthenticated."
...
paths:
...
  /drugs/search:
    get:
      tags: [Prescriptions]
      summary: Drug database search
      description: Searches the curated Indian drug database by name, generic or brand. Supports specialty-specific filtering.
      operationId: drugSearch
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            example: "methotrexate"
        - name: specialty
          in: query
          schema:
            type: string
            example: "dermatology"
      responses:
        '200':
          description: Matching drugs
          content:
            application/json:
              schema:
                type: object
                properties:
                  drugs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                          example: "Methotrexate"
                        generic_name:
                          type: string
                          example: "Methotrexate"
                        category:
                          type: string
                          example: "Immunosuppressant"
                        common_doses:
                          type: array
                          items:
                            type: string
                          example: ["7.5mg weekly", "15mg weekly"]
                        route:
                          type: string
                          example: "oral"
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ═══════════════════════════════════════════════════════════════════════════
  # BILLING & INVOICES
  # ═══════════════════════════════════════════════════════════════════════════
...